fix: migrate media from screenshots-pending when approved from the review tool
